Added a max number of items that can be displayed in a menu. Fixed some of the layout bugs in the menu. Changed the data structures a bit.

+++ EMBRACE_Client/PieContextualMenuItem.h
@@ -17,6 +17,6 @@
 @property (nonatomic, strong) MenuItemDataSource *data;
 
 - (id)initWithFrameAndData:(CGRect)frame :(float)angleFromCenter :(MenuItemDataSource*)itemData;
-- (UIImage *)scaleImagetoResolution:(UIImage*)image :(int)resolution;
+- (UIImage *)scaleImagetoResolution:(UIImage*)image :(float)resolution;
 
 @end
